+++ apps/workflow-server/gcc-workflow-server-facade/gcc-restclient-facade-default/src/test/java/com/coremedia/labs/translation/gcc/facade/def/DefaultGCExchangeFacadeTest.java
@@ -155,7 +155,7 @@ class DefaultGCExchangeFacadeTest {
     when(gcExchange.uploadContent(any())).thenThrow(RuntimeException.class);
 
     GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange);
-    assertThatThrownBy(() -> facade.uploadContent(expectedFileName, new ByteArrayResource(new byte[]{42}), null))
+    assertThatThrownBy(() -> facade.uploadContent(expectedFileName, new ByteArrayResource(new byte[]{(byte) 42}), null))
             .isInstanceOf(GCFacadeCommunicationException.class)
             .hasCauseInstanceOf(RuntimeException.class)
             .hasMessageContaining(expectedFileName);
